feat: more security, move service context, add cron, log

+++ src/backend/services/account/business/delete_account.go
@@ -2,29 +2,32 @@ package business
 
 import (
 	"context"
+	"errors"
+	"tradeplay/common"
 	"tradeplay/services/account/entity"
-
-	"github.com/DatLe328/service-context/core"
 )
 
-func (biz *business) DeleteAccount(ctx context.Context, id int) error {
+func (biz *business) DeleteAccount(ctx context.Context, id int32) error {
 	oldData, err := biz.accountRepo.GetAccountByID(ctx, id)
 	if err != nil {
-		return core.ErrCannotGetEntity(entity.Account{}.TableName(), err)
+		return common.ErrCannotGetEntity(entity.Account{}.TableName(), err)
+	}
+
+	if oldData.Status == entity.AccountStatusSold {
+		return common.ErrInvalidRequest(errors.New("cannot delete a sold account"))
 	}
 
 	if err := biz.orderRepo.DeleteOrdersByAccountId(ctx, id); err != nil {
-		return core.ErrInternal(err)
+		return common.ErrInternal(err)
 	}
 
 	if len(oldData.Images) > 0 {
 		if err := biz.uploadComponent.DeleteFiles(ctx, oldData.Images); err != nil {
-			return core.ErrInternal(err)
 		}
 	}
 
 	if err := biz.accountRepo.DeleteAccount(ctx, id); err != nil {
-		return core.ErrInternal(err)
+		return common.ErrInternal(err)
 	}
 
 	return nil
